feat: 두께 설정에 따른 preview 함수 viewmodel로 분리
