+++ app/src/androidTest/java/com/example/blurit/edit/EditViewModelTest.kt
@@ -1,11 +1,8 @@
 import android.graphics.Bitmap
-import android.graphics.Rect
 import android.widget.ImageView
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.example.blurit.edit.EditViewModel
 import com.google.mlkit.vision.face.Face
-import com.google.mlkit.vision.face.FaceContour
-import com.google.mlkit.vision.face.FaceLandmark
 import junit.framework.TestCase.assertEquals
 import junit.framework.TestCase.assertNotNull
 import org.junit.Assert.assertNotEquals
@@ -31,16 +28,13 @@ class EditViewModelTest {
 
     @Test
     fun initBitmapTest() {
-        // Given
         val originalBitmap = Bitmap.createBitmap(100, 120, Bitmap.Config.ARGB_8888)
         val imageViewWidth = 200
         val ratio = originalBitmap.width.toFloat() / originalBitmap.height.toFloat()
         val imageViewHeight = (imageViewWidth / ratio).toInt()
 
-        // When
         viewModel.initBitmap(originalBitmap, imageViewWidth)
 
-        // Then
         val initializedOriginalCanvas = viewModel.originalCanvas.value
         val initializedBlurCanvas = viewModel.blurCanvas.value
         val initializedThicknessCanvas = viewModel.thicknessCanvas.value
@@ -61,16 +55,13 @@ class EditViewModelTest {
 
     @Test
     fun applyAutoMosaicShouldThrowNPEWhenBoundingBoxIsNull() {
-        // Given
         val originalBitmap = Bitmap.createBitmap(100, 100, Bitmap.Config.ARGB_8888)
         viewModel.initBitmap(originalBitmap, 100)
 
-        // Mocking Face object with null boundingBox
         val face = Mockito.mock(Face::class.java)
         `when`(face.boundingBox).thenReturn(null)
         val faces = listOf(face)
 
-        // When & Then: Verify NullPointerException is thrown
         assertThrows(NullPointerException::class.java) {
             viewModel.applyAutoMosaic(faces, 5)
         }
@@ -78,25 +69,20 @@ class EditViewModelTest {
 
     @Test
     fun applyManualMosaicTest() {
-        // Given
         val originalBitmap = Bitmap.createBitmap(100, 100, Bitmap.Config.ARGB_8888)
         viewModel.initBitmap(originalBitmap, 100)
 
-        // Mock an ImageView and set up test coordinates and parameters
         val mockImageView = mock(ImageView::class.java)
         val touchX = 50
         val touchY = 50
         val thickness = 10
         val blur = 5
 
-        // Set initial blurCanvas to observe changes
         val initialBlurCanvas = viewModel.blurCanvas.value
         assertNotNull("blurCanvas should be initialized", initialBlurCanvas)
 
-        // When
         viewModel.applyManualMosaic(mockImageView, touchX, touchY, thickness, blur)
 
-        // Then
         val updatedBlurCanvas = viewModel.blurCanvas.value
         assertNotNull("blurCanvas should be updated after applyManualMosaic", updatedBlurCanvas)
